Update Volume_calc.py
20230072 이예나 중앙대학교 오픈소스sw와 파이썬 프로그래밍 1주차 과제 박스의 부피 구하기

+++ Volume_calc.py
@@ -1,6 +1,6 @@
-widht = int(input("widht: "))
-length = int(input("length: "))
-high = int(input("high: "))
+widht = float(input("widht: "))
+length = float(input("length: "))
+high = float(input("high: "))
 
 volume = widht * length * high 
 print("The volume of the box is", volume)
